src/main/java/Ticket.java
public class Ticket implements Comparable<Ticket> {
    protected int id;
    protected int price;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public int getPrice() {
        return price;
    }

    public void setPrice(int price) {
        this.price = price;
    }

    public String getFrom() {
        return from;
    }

    public void setFrom(String from) {
        this.from = from;
    }

    public String getTo() {
        return to;
    }

    public void setTo(String to) {
        this.to = to;
    }

    public int getMinutes() {
        return minutes;
    }

    public void setMinutes(int minutes) {
        this.minutes = minutes;
    }

    protected String from;
    protected String to;
    protected int minutes;

    public Ticket(int id, int price, String from, String to, int minutes) {
        this.id = id;
        this.price = price;
        this.from = from;
        this.to = to;
        this.minutes = minutes;

    }

    @Override
    public int compareTo(Ticket o) {
        if (this.price < o.price) {
            return -1;
        } else if (this.price > o.price) {
            return 1;
        }
        return 0;
    }
}
